added some functions
Issue with totalSacksPerGame().  Initially was returning NaN in return arr when a game was not played by all players.  Created workaround to avoid games not played by all players but as a result, the returning Arr does not include totals for games that were not played by all.  However no NaN in the returning array but missing info.  Need to devise solution to this issue